Create template for service proposal  (#8)
* Create template for service proposal

* Update service-proposal.md from review

These changes clarify when more explicit information must be included in the service proposal issue. Any service, whether Bluetooth SIG standardized or standardized by another organization, can be summarized by attaching or providing a link to download an existing comprehensive service specification. If such a document is not available, this information must be detailed in the issue itself.

Co-authored-by: George Beckstein <donbeckstein@live.com>